Move capabilities to its own file
