+++ pkg/security/probe/probe.go
@@ -56,15 +56,6 @@ type Probe struct {
 	syscallMonitor   *SyscallMonitor
 }
 
-// Capability represents the type of values we are able to filter kernel side
-type Capability struct {
-	PolicyFlags     PolicyFlag
-	FieldValueTypes eval.FieldValueType
-}
-
-// Capabilities represents the filtering capabilities for a set of fields
-type Capabilities map[eval.Field]Capability
-
 // HookPoint represents
 type HookPoint struct {
 	Name            string
@@ -314,40 +305,6 @@ var allHookPoints = []*HookPoint{
 			"link": {},
 		},
 	},
-}
-
-// GetFlags returns the policy flags for the set of capabilities
-func (caps Capabilities) GetFlags() PolicyFlag {
-	var flags PolicyFlag
-	for _, cap := range caps {
-		flags |= cap.PolicyFlags
-	}
-	return flags
-}
-
-// GetFields returns the fields associated with a set of capabilities
-func (caps Capabilities) GetFields() []eval.Field {
-	var fields []eval.Field
-
-	for field := range caps {
-		fields = append(fields, field)
-	}
-
-	return fields
-}
-
-// GetFieldCapabilities returns the field capabilities for a set of capabilities
-func (caps Capabilities) GetFieldCapabilities() rules.FieldCapabilities {
-	var fcs rules.FieldCapabilities
-
-	for field, cap := range caps {
-		fcs = append(fcs, rules.FieldCapability{
-			Field: field,
-			Types: cap.FieldValueTypes,
-		})
-	}
-
-	return fcs
 }
 
 // NewRuleSet returns a new rule set
